missed the tests

+++ src/test/scala/org/ialab/hmm/HMMTest.scala
@@ -11,8 +11,8 @@ import org.ialab.hmm.FloatMatrix.ConstantMatrix
 class HMMTest extends FlatSpec with PrivateMethodTester {
 
   it should "find the state sequence 0, 1 - 0, 0 and 1, 1 for 3 different observations" in {
-    val hmm = new HMM(2, 2, Array[Double](0.5f, 0.5f), ConstantMatrix(2, 2, 0.5f),
-      FloatMatrix(Array[Double](1, 0, 0, 1), 2))
+    val hmm = new HMM(2, 2, Array[Float](0.5f, 0.5f), ConstantMatrix(2, 2, 0.5f),
+      FloatMatrix(Array[Float](1, 0, 0, 1), 2))
 
     assert(hmm.viterbi(Array[Int] (0, 1))._2.sameElements(Array[Float](0,1)))
 
@@ -23,26 +23,26 @@ class HMMTest extends FlatSpec with PrivateMethodTester {
   }
 
   it should "find the state sequence 0, 1, 1" in {
-    val hmm = new HMM(3, 2, Array[Double](0.5f, 0.5f, 0), FloatMatrix(Array[Double](0.5f, 0.5f, 0.5f, 0, 0.5f, 0.5f, 0.5f, 0.5f, 0.5f), 3),
-      FloatMatrix(Array[Double](1, 0, 0, 1, 1, 0), 2))
+    val hmm = new HMM(3, 2, Array[Float](0.5f, 0.5f, 0), FloatMatrix(Array[Float](0.5f, 0.5f, 0.5f, 0, 0.5f, 0.5f, 0.5f, 0.5f, 0.5f), 3),
+      FloatMatrix(Array[Float](1, 0, 0, 1, 1, 0), 2))
 
-    assert(hmm.viterbi(Array[Int] (0, 1, 1))._2.sameElements(Array[Double](0,1,1)))
-    assert(hmm.viterbi(Array[Int] (0, 1, 0))._2.sameElements(Array[Double](0,1,2)))
+    assert(hmm.viterbi(Array[Int] (0, 1, 1))._2.sameElements(Array[Float](0,1,1)))
+    assert(hmm.viterbi(Array[Int] (0, 1, 0))._2.sameElements(Array[Float](0,1,2)))
   }
 
   it should "Learn a simple two state where the first states emits 0 and the second emits 1" in {
     val observations = Array[Int](0,0,0,1,0,0,1,0,0,0,0,1,0,0,0,0,0,1)
-    val hmm = new HMM(2, 2, Array[Double](0.9f, 0.1f), ConstantMatrix(2, 2, 0.5f),
-      FloatMatrix(Array[Double](0.7f, 0.3f, 0.3f, 0.7f), 2))
+    val hmm = new HMM(2, 2, Array[Float](0.9f, 0.1f), ConstantMatrix(2, 2, 0.5f),
+      FloatMatrix(Array[Float](0.7f, 0.3f, 0.3f, 0.7f), 2))
 
-    assert(hmm.viterbi(Array[Int] (0,0,0,0,1))._2.sameElements(Array[Double](0,0,0,0,1)))
-    assert(hmm.viterbi(Array[Int] (0,0,0,1,1))._2.sameElements(Array[Double](0,0,0,1,1)))
+    assert(hmm.viterbi(Array[Int] (0,0,0,0,1))._2.sameElements(Array[Float](0,0,0,0,1)))
+    assert(hmm.viterbi(Array[Int] (0,0,0,1,1))._2.sameElements(Array[Float](0,0,0,1,1)))
   }
 
   it should "It should be unlikely to see a chain with many 1" in {
     val observations = Array[Int](0,0,0,1,0,0,1,0,0,0,0,1,0,0,0,0,0,1)
-    val hmm = new HMM(2, 2, Array[Double](0.5f, 0.5f), ConstantMatrix(2, 2, 0.5f),
-      FloatMatrix(Array[Double](0.7f, 0.3f, 0.3f, 0.7f), 2))
+    val hmm = new HMM(2, 2, Array[Float](0.5f, 0.5f), ConstantMatrix(2, 2, 0.5f),
+      FloatMatrix(Array[Float](0.7f, 0.3f, 0.3f, 0.7f), 2))
 
     hmm.learn(observations, 20)
 
@@ -59,10 +59,10 @@ class HMMTest extends FlatSpec with PrivateMethodTester {
 
   it should "Learn a simple three state where the first state emits 0, the second emits 1 and the third emits 0, no return from second state to first one" in {
     val observations = Array[Int](0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)
-    val hmm = new HMM(3, 2, Array[Double](1f, 0f, 0f), FloatMatrix(Array[Double](0.5f, 0.5f, 0f,
+    val hmm = new HMM(3, 2, Array[Float](1f, 0f, 0f), FloatMatrix(Array[Float](0.5f, 0.5f, 0f,
                                                                                 0f, 0.5f, 0.5f,
                                                                                 0f, 0f, 1f), 3),
-      FloatMatrix(Array[Double](1f, 0f,
+      FloatMatrix(Array[Float](1f, 0f,
                               0f, 1f,
                               1f, 0f), 2))
 
@@ -70,7 +70,7 @@ class HMMTest extends FlatSpec with PrivateMethodTester {
     hmm.learn(observations, 20)
 
 
-    assert(hmm.viterbi(Array[Int] (0,0,0,0,1,0))._2.sameElements(Array[Double](0,0,0,0,1,2)))
+    assert(hmm.viterbi(Array[Int] (0,0,0,0,1,0))._2.sameElements(Array[Float](0,0,0,0,1,2)))
 
     assert(hmm.viterbi(Array[Int] (0,0,0,0,1))._1 > hmm.viterbi(Array[Int] (1,1,1,1,0))._1)
   }
